% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{auc_gradient}
\alias{auc_gradient}
\title{Gradient of the ramp pseudo AUC}
\usage{
auc_gradient(X, pseus, rampfunc = smoothramp,
  d.rampfunc = d.smoothramp)
}
\arguments{
\item{X}{numeric vector of predictions}

\item{pseus}{matrix of pseudo observations}

\item{rampfunc}{Ramp function}

\item{d.rampfunc}{first derivative of ramp function}
}
\description{
Gradient of the AUC using the specified ramp function and the first derivative of the ramp.
Returns the gradient with respect to the prediction vector. This is for use with xgboost.
}
