% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{auc_hessian}
\alias{auc_hessian}
\title{Hessian of the ramp pseudo AUC}
\usage{
auc_hessian(X, pseus, rampfunc = smoothramp, d.rampfunc = d.smoothramp,
  d2.rampfunc = d2.smoothramp)
}
\arguments{
\item{X}{numeric vector of predictions}

\item{pseus}{matrix of pseudo observations}

\item{rampfunc}{Ramp function}

\item{d.rampfunc}{first derivative of ramp function}

\item{d2.rampfunc}{second derivative of ramp function (only the diagonal terms are needed)}
}
\description{
Hessian of the AUC using the specified ramp function and the first and second derivatives
of the ramp. Returns the hessian with respect to the prediction vector. This is mainly
for use with xgboost.
}
