% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{calc_auc}
\alias{calc_auc}
\title{Compute AUROC using the trapezoidal rule}
\usage{
calc_auc(fpf, tpf)
}
\arguments{
\item{fpf}{Vector of false positive fractions/1 - specificity}

\item{tpf}{Vector of true positive fractions/sensitivy}
}
\description{
Compute AUROC using the trapezoidal rule
}
