% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{calc_roc}
\alias{calc_roc}
\title{Compute ROC curve based on pseudo observations}
\usage{
calc_roc(X, pseus, ramp = function(X.vect, c.scal) as.numeric(X.vect >
  c.scal))
}
\arguments{
\item{X}{the predicted pseudo observations}

\item{pseus}{A matrix or data frame, pseudo obs of interest in the first column}

\item{ramp}{A ramp function, first argument is a vector, and second is a scalar}
}
\value{
A data frame with columns for the false and true positive fractions
}
\description{
Given a predictor X and a matrix of pseudo observations, with one
column per cause, and only the pseudo observations for a fixed time
computes the true and false positive fractions for each unique value
of the predictor.
}
