% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp-functions.R
\name{d2.smoothramp}
\alias{d2.smoothramp}
\title{Second derivative of sigmoid ramp function}
\usage{
d2.smoothramp(X.vect, c.scal, sigma = 0.05)
}
\description{
Second derivative of sigmoid ramp function
}
