% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-pseudo.R
\name{make_pseudo_matrix}
\alias{make_pseudo_matrix}
\title{Make pseudo observation evaluation matrix}
\usage{
make_pseudo_matrix(pseudo, time = 1)
}
\arguments{
\item{pseudo}{an object as returned from pseudoci, pseudosurv, or similar}

\item{time}{index of the time point of interest}
}
\description{
Given a pseudo object, and a time of interest,
stacks in columns the pseudo observations of the
different causes at that time. Result is suitable for use
with the ROC and AUC functions.
}
