% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-pseudo.R
\name{make_pseudo_predictor}
\alias{make_pseudo_predictor}
\title{Make pseudo observation predictor data frame}
\usage{
make_pseudo_predictor(pseudo, cause = 1)
}
\arguments{
\item{pseudo}{an object as returned from pseudoci, pseudosurv, or similar}

\item{cause}{index of the cause of interest}
}
\description{
Given a pseudo object, stacks in rows the pseudo observations for
a fixed cause over the different times at which the pseudo
observations were calculated, along with the time at which they
were calculated.
}
