% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{multivar_auc_gradient}
\alias{multivar_auc_gradient}
\title{Gradient of AUC with respect to vector of predictors}
\usage{
multivar_auc_gradient(X.wide, pseus, rampfunc = smoothramp,
  d.rampfunc = d.smoothramp)
}
\arguments{
\item{X.wide}{matrix of predictions}

\item{pseus}{matrix of pseudo observations}

\item{rampfunc}{ramp function}

\item{d.rampfunc}{first derivative of the ramp function with respect to the coefficients}
}
\description{
Given a vector of numeric predictors for each subject, this calculates the gradient
of the pseudo-AUC with respect to the coefficients for the linear combination of
the predictors. This can be used for AUC regression, wherein the predictors are independent
biomarkers, and also with SuperLearner, wherein the predictors are the prediction
results of a library of prediction algorithms.
}
