% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-for-xgboost.R
\name{pseudoauc_objective}
\alias{pseudoauc_objective}
\title{Pseudo observation AUC objective function}
\usage{
pseudoauc_objective(preds, dtrain)
}
\arguments{
\item{preds}{Vector of predictions}

\item{dtrain}{An object of class xgb.Dmatrix, with additional attributes otherc and tdex}
}
\description{
Designed for use with xgboost, given a vector of predictions and
a training matrix, returns the gradient and hessian so that it can
be optimized by xgboost
}
