% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{ramp_auc}
\alias{ramp_auc}
\title{Ramp auc calculation}
\usage{
ramp_auc(X, pseus, rampfunc = stepramp)
}
\arguments{
\item{X}{numeric vector of predictions}

\item{pseus}{matrix of pseudo observations}

\item{rampfunc}{Ramp function}
}
\description{
Calculates the approximate pseudo AUC using the specified ramp function
to be used as the indicator. Using the indicator function as the ramp will
result in the empirical AUC.
}
